+++ watchu/watchu/src/main/java/com/wc/watchu/repository/ContentsRepository.java
@@ -7,11 +7,11 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ContentsRepository extends JpaRepository<ContentsEntity, Long>
+public interface ContentsRepository extends JpaRepository<ContentsEntity, Integer>
 {
     // 카테고리로 콘텐츠를 조회 (예: MOVIE, SERIES)
     List<ContentsEntity> findByContentsCategoryOrderByContentsIdDesc(String contentsCategory);
 
     // 콘텐츠 ID로 단일 콘텐츠 조회
-    ContentsEntity findByContentsId(Long contentsId);
+    ContentsEntity findByContentsId(int contentsId);
 }
